add new comment

+++ backend/project/src/main/java/com/example/project/dto/CommentDTO.java
@@ -9,7 +9,7 @@ public class CommentDTO {
     private int rating;
     private LocalDate updateDate;
     private UserDTO userName;
-    private int postId;
+    private Long postId;
 
     public int getId() {
         return id;
@@ -43,11 +43,11 @@ public class CommentDTO {
         this.updateDate = updateDate;
     }
 
-    public int getPostId() {
+    public Long getPostId() {
         return postId;
     }
 
-    public void setPostId(int postId) {
+    public void setPostId(Long postId) {
         this.postId = postId;
     }
